+++ crates/apps/looper/looper-processor/src/controllers/mod.rs
@@ -21,6 +21,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 pub mod audio_io_settings_controller;
+pub mod audio_state_controller;
 pub mod autosave_controller;
 pub mod events_controller;
 pub mod load_project_controller;
